ebook-management-system/backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-MySQLdb==1.0.1
PyJWT==2.8.0
Werkzeug==3.0.3
mysqlclient==2.2.4
